foosrefmanagement/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-better-admin-arrayfield==1.4.2
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
